% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PML.R
\name{PML}
\alias{PML}
\title{Load the 24 prices by hour of a node in a specific day}
\usage{
PML(node = "01PLO-115", syst = "SIN", market = "MDA", fecha = "2017/07/11")
}
\arguments{
\item{node}{The node of the electric system; defaults to 01PLO-115}

\item{syst}{One of the three electrical systems of Mexico: SIN, BCA, or BCS; defaults to SIN}

\item{market}{One of the two markets in Mexico: MDA or MTR; defaults to MDA}

\item{fecha}{The date; defaults to 2017/07/11}
}
\value{
Dataframe with the 24 hours and a price for each hour
}
\description{
Load the 24 prices by hour of a node in a specific day
}
\examples{
PML()
}
